pyproject.toml
[project]
name = "number-recognition"
version = "0.1.0"
description = "Automatic license plate recognition pipeline"
authors = [{name = "Auto Developer", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "pydantic-settings>=2.2.1",
    "pydantic>=2.7.1",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.8.0",
    "aiofiles>=23.2.1",
    "prometheus-client>=0.20.0"
]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
